// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
      crops_dataset: {
        Row: {
          created_at: string
          crop: string
          fertilizer: number
          humidity: number
          id: string
          rainfall: number
          soil_ph: number
          temperature: number
          yield: number
        }
        Insert: {
          created_at?: string
          crop: string
          fertilizer: number
          humidity: number
          id?: string
          rainfall: number
          soil_ph: number
          temperature: number
          yield: number
        }
        Update: {
          created_at?: string
          crop?: string
          fertilizer?: number
          humidity?: number
          id?: string
          rainfall?: number
          soil_ph?: number
          temperature?: number
          yield?: number
        }
        Relationships: []
      }
      model_metrics: {
        Row: {
          created_at: string
          id: string
          mae: number
          model_name: string
          r2_score: number
          rmse: number
          training_date: string
        }
        Insert: {
          created_at?: string
          id?: string
          mae: number
          model_name: string
          r2_score: number
          rmse: number
          training_date?: string
        }
        Update: {
          created_at?: string
          id?: string
          mae?: number
          model_name?: string
          r2_score?: number
          rmse?: number
          training_date?: string
        }
        Relationships: []
      }
      predictions: {
        Row: {
          best_model: string
          created_at: string
          fertilizer: number
          humidity: number
          id: string
          predicted_crop: string
          predicted_yield_lr: number
          predicted_yield_rf: number
          rainfall: number
          soil_ph: number
          temperature: number
        }
        Insert: {
          best_model: string
          created_at?: string
          fertilizer: number
          humidity: number
          id?: string
          predicted_crop: string
          predicted_yield_lr: number
          predicted_yield_rf: number
          rainfall: number
          soil_ph: number
          temperature: number
        }
        Update: {
          best_model?: string
          created_at?: string
          fertilizer?: number
          humidity?: number
          id?: string
          predicted_crop?: string
          predicted_yield_lr?: number
          predicted_yield_rf?: number
          rainfall?: number
          soil_ph?: number
          temperature?: number
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never